requirements.txt
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
numpy<2.0
pandas==2.0.3
Pillow==10.0.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
opencv-python==4.8.0.76
albumentations==1.3.1
tqdm==4.65.0
PyYAML==6.0.2

# For Streamlit app
streamlit==1.28.0
plotly==5.17.0

# For training monitoring
tensorboard==2.14.0
wandb==0.15.12

# Optional but recommended
einops==0.7.0

# For AI Agent
pydantic-ai
openai
nest-asyncio